Extract chat box from base chat
